avoid int overflow in random number generation
--HG--
extra : convert_revision : 18642a2e6c8a7d4409800b0b0da83448ec8551a1

+++ branches/ompl-isucan/src/ompl/util/src/RandomNumbers.cpp
@@ -76,7 +76,7 @@ double ompl::RNG::halfNormalReal(double r_min, double r_max, double focus)
 
 int ompl::RNG::halfNormalInt(int r_min, int r_max, double focus)
 {
-    return (int)halfNormalReal(r_min, r_max + (1.0 - 1e-9), focus);
+    return (int)halfNormalReal((double)r_min, (double)r_max + (1.0 - 1e-9), focus);
 }
 
 // From: "Uniform Random Rotations", Ken Shoemake, Graphics Gems III,
